comit 23-06-22

+++ demo/src/main/java/com/demo/repository/UsuarioRespository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.demo.model.Usuario;
 
-public interface UsuarioRespository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRespository extends JpaRepository<Usuario, Integer> {
 
 }
